Move InvokeResponse and TypedInvokeResponse to bot-schema
