+++ libraries/bot-builder/src/main/java/com/microsoft/bot/builder/InvokeResponse.java
@@ -9,18 +9,8 @@ package com.microsoft.bot.builder;
  * the resulting POST. The Body of the resulting POST will be the JSON
  * Serialized content from the Body property.
  */
-public class InvokeResponse {
-
-    /**
-     * The POST that is generated in response to the incoming Invoke Activity will
-     * have the HTTP Status code specified by this field.
-     */
-    private int status;
-    /**
-     * The POST that is generated in response to the incoming Invoke Activity will
-     * have a body generated by JSON serializing the object in the Body field.
-     */
-    private Object body;
+@Deprecated // Use the one of bot-schema
+public class InvokeResponse extends com.microsoft.bot.schema.InvokeResponse {
 
     /**
      * Initializes new instance of InvokeResponse.
@@ -29,53 +19,6 @@ public class InvokeResponse {
      * @param withBody   The invoke response body.
      */
     public InvokeResponse(int withStatus, Object withBody) {
-        status = withStatus;
-        body = withBody;
+        super(withStatus, withBody);
     }
-
-    /**
-     * Gets the HTTP status code for the response.
-     *
-     * @return The HTTP status code.
-     */
-    public int getStatus() {
-        return status;
-    }
-
-    /**
-     * Sets the HTTP status code for the response.
-     *
-     * @param withStatus The HTTP status code.
-     */
-    public void setStatus(int withStatus) {
-        this.status = withStatus;
-    }
-
-    /**
-     * Gets the body content for the response.
-     *
-     * @return The body content.
-     */
-    public Object getBody() {
-        return body;
-    }
-
-    /**
-     * Sets the body content for the response.
-     *
-     * @param withBody The body content.
-     */
-    public void setBody(Object withBody) {
-        body = withBody;
-    }
-
-    /**
-     * Returns if the status of the request was successful.
-     * @return True if the status code is successful, false if not.
-     */
-    @SuppressWarnings("MagicNumber")
-    public boolean getIsSuccessStatusCode() {
-        return status >= 200 && status <= 299;
-    }
-
 }
