+++ libraries/bot-schema/src/main/java/com/microsoft/bot/schema/TypedInvokeResponse.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.bot.builder;
+package com.microsoft.bot.schema;
 
 /**
  * Tuple class containing an HTTP Status Code and a JSON Serializable object.
@@ -10,8 +10,7 @@ package com.microsoft.bot.builder;
  * Serialized content from the Body property.
  * @param <T> The type for the body of the TypedInvokeResponse.
  */
-@Deprecated // Use the one of bot-schema
-public class TypedInvokeResponse<T> extends com.microsoft.bot.schema.TypedInvokeResponse {
+public class TypedInvokeResponse<T> extends InvokeResponse {
 
     /**
      * Initializes new instance of InvokeResponse.
@@ -21,5 +20,22 @@ public class TypedInvokeResponse<T> extends com.microsoft.bot.schema.TypedInvoke
      */
     public TypedInvokeResponse(int withStatus, T withBody) {
         super(withStatus, withBody);
+    }
+
+    /**
+     * Sets the body with a typed value.
+     * @param withBody the typed value to set the body to.
+     */
+    public void setTypedBody(T withBody) {
+        super.setBody(withBody);
+    }
+
+    /**
+     * Gets the body content for the response.
+     *
+     * @return The body content.
+     */
+    public T getTypedBody() {
+        return (T) super.getBody();
     }
 }
